Allow to create a new commit by amending to an existing one.
