+++ crates/but-core/tests/core/diff/worktree_changes.rs
@@ -236,7 +236,7 @@ fn added_in_unborn() -> Result<()> {
 
 #[test]
 fn sparse() -> Result<()> {
-    let repo = repo("sparse")?;
+    let repo = repo_in("sparse", "non-cone")?;
     let err = diff::worktree_changes(&repo).unwrap_err();
     assert!(
         err.to_string().contains("sparse"),
@@ -1233,16 +1233,20 @@ fn unified_diffs(
         .collect()
 }
 
-pub fn repo(fixture_name: &str) -> anyhow::Result<gix::Repository> {
-    let root = gix_testtools::scripted_fixture_read_only("worktree-changes.sh")
+pub fn repo_in(fixture_name: &str, name: &str) -> anyhow::Result<gix::Repository> {
+    let root = gix_testtools::scripted_fixture_read_only(format!("{}.sh", fixture_name))
         .map_err(anyhow::Error::from_boxed)?;
-    let worktree_root = root.join(fixture_name);
-    Ok(gix::open(worktree_root)?)
+    let worktree_root = root.join(name);
+    Ok(gix::open_opts(
+        worktree_root,
+        gix::open::Options::isolated(),
+    )?)
+}
+
+pub fn repo(fixture_name: &str) -> anyhow::Result<gix::Repository> {
+    repo_in("worktree-changes", fixture_name)
 }
 
 pub fn repo_unix(fixture_name: &str) -> anyhow::Result<gix::Repository> {
-    let root = gix_testtools::scripted_fixture_read_only("worktree-changes-unix.sh")
-        .map_err(anyhow::Error::from_boxed)?;
-    let worktree_root = root.join(fixture_name);
-    Ok(gix::open(worktree_root)?)
+    repo_in("worktree-changes-unix", fixture_name)
 }
