+++ crates/but-core/tests/fixtures/worktree-changes.sh
@@ -194,20 +194,3 @@ cat <<EOF >>.git/config
 	textconv = "shift; echo ho"
 EOF
 )
-
-
-git init -q sparse
-(cd sparse
-  touch a b
-  mkdir c1
-  (cd c1 && touch a b && mkdir c2 && cd c2 && touch a b)
-  (cd c1 && mkdir c3 && cd c3 && touch a b)
-  mkdir d
-  (cd d && touch a b && mkdir c4 && cd c4 && touch a b c5)
-
-  git add .
-  git commit -m "init"
-
-  git sparse-checkout set c1/c2 --sparse-index
-)
-
