+++ crates/but-workspace/tests/workspace/commit_engine/mod.rs
@@ -1,0 +1,3 @@
+mod amend_commit;
+mod new_commit;
+mod utils;
